==> Preprocessing/6_combining_tweet_and_cohort.ipynb <==
# Combining tweet and timezone datasets with cohort

import pandas as pd

# Cohort files
d_cohort = pd.read_csv('../Data/Combined_sets/depression_processed_cohort_tz.tsv', sep='\t', index_col="tweet_id")
a_cohort = pd.read_csv('../Data/Combined_sets/anxiety_processed_cohort_tz.tsv', sep='\t', index_col="tweet_id")

# Tweet files

d_tweets = pd.read_csv('../Data/Combined_sets/d_processed_tweets.tsv', sep='\t', index_col="tweet_id")
a_tweets = pd.read_csv('../Data/combined_sets/a_processed_tweets.tsv', sep='\t', index_col="tweet_id")

columns_to_drop = ['Unnamed: 0']
d_cohort = d_cohort.drop(columns = columns_to_drop)
a_cohort = a_cohort.drop(columns = columns_to_drop)


Drop preprocessed tweets column

d_tweets = d_tweets.drop(['edited_text'], axis=1)
a_tweets = a_tweets.drop(['edited_text'], axis=1)

Join the two

depression = d_cohort.join(d_tweets)
anxiety = a_cohort.join(a_tweets)

## Save these datasets into one

depression.to_csv('../Data/combined_sets/d_combined_cohort_tz.tsv', sep='\t')
anxiety.to_csv('../Data/combined_sets/a_combined_cohort_tz.tsv', sep='\t')
